pyproject.toml
[project]
name = "test-up-sound"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.27.0",
    "python-dotenv>=1.2.2",
    "yandex-music>=3.0.0",
]
